ai-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.25.0
python-dotenv>=1.0.1
pydantic>=2.7.0
httpx>=0.27.0
